Function jsonwebtokens::raw::split_token

pub fn split_token<'a>(token: &'a str) -> Result<TokenSlices<'a>, Error>

Splits a token that's in the form "HEADER.CLAIMS.SIGNATURE" into useful constituent parts for further parsing and validation.

For example:

let token = "HEADER.CLAIMS.SIGNATURE";
let TokenSlices {message, signature, header, claims } = raw::split_token(token)?;
println!("message: {}", message);
println!("signature: {}", signature);
println!("header: {}", header);
println!("claims: {}", claims);

will output:

message: HEADER.CLAIMS
signature: SIGNATURE
header: HEADER
claims: CLAIMS

After splitting a token, it can be further processed by using raw::verify_signature_only to check the token's signature, then raw::decode_json_token_slice can be used to decode the header and/or the claims, and finally the Verifier::verify_claims_only api can be used to check the claims.